+++ src/script.cpp
@@ -100,11 +100,11 @@ const unordered_map<string, unsigned int> Script::findKeyword(const string & key
 
 	for(const auto & [scriptFileId, scriptFile] : _scriptFiles)
 	{
-		for(unsigned int lineNumber = 0; lineNumber < scriptFile->getLines().size(); lineNumber++)
+		for(unsigned int lineNumber = 0; lineNumber < static_cast<unsigned int>(scriptFile->getLines().size()); lineNumber++)
 		{
 			const auto line = scriptFile->getLines()[lineNumber];
 
-			for(unsigned int index = 0; index < line.size(); index++)
+			for(unsigned int index = 0; index < static_cast<unsigned int>(line.size()); index++)
 			{
 				if(line.substr(index, keyword.size()) == keyword)
 				{
