+++ src/texture_buffer.cpp
@@ -47,7 +47,7 @@ TextureBuffer::TextureBuffer(const array<shared_ptr<Image>, 6> & images)
 		}
 	}
 
-	for(unsigned int index = 0; index < images.size(); index++)
+	for(unsigned int index = 0; index < static_cast<unsigned int>(images.size()); index++)
 	{
 		const auto cubeIndex = (GL_TEXTURE_CUBE_MAP_POSITIVE_X + static_cast<unsigned int>(index));
